+++ HISTORY.md
@@ -1,7 +1,7 @@
 History of features added to LiChess Tools in time (reversed order)
 
-05 Jun
-  - paste images in chat directly
+06 Jun
+  - paste images in chat/forum
 
 04 Jun
   - General tool for extra notifications